cache added

+++ src/main/java/com/test/bus_route_challenge/model/DirectBusRouteResponse.java
@@ -9,8 +9,8 @@ import lombok.Setter;
 @AllArgsConstructor
 public class DirectBusRouteResponse {
 
-    private Long dep_sid;
-    private Long arr_sid;
+    private Integer dep_sid;
+    private Integer arr_sid;
     private boolean direct_bus_route;
 
 }
